modulesync 7.3.0-8-gd083aee

+++ lib/puppet/type/cpanm.rb
@@ -1,19 +1,17 @@
+# frozen_string_literal: true
+
 Puppet::Type.newtype(:cpanm) do
-  @doc = "Manage CPAN modules with cpanminus"
+  @doc = 'Manage CPAN modules with cpanminus'
   ensurable do
     attr_accessor :latest
 
     defaultto :installed
     newvalue(:absent) do
-      if provider.exists?
-        provider.destroy
-      end
+      provider.destroy if provider.exists?
     end
 
     newvalue(:present) do
-      unless provider.exists?
-        provider.create
-      end
+      provider.create unless provider.exists?
     end
     aliasvalue(:installed, :present)
 
@@ -26,19 +24,18 @@ Puppet::Type.newtype(:cpanm) do
 
     def insync?(is)
       @should.each do |should|
-      case should
-      when :present
-        return true if is == :present
-      when :latest
-        return false if is == :absent
-        return true if provider.latest?
-      when :absent
-        return true if is == :absent
-      end
+        case should
+        when :present
+          return true if is == :present
+        when :latest
+          return false if is == :absent
+          return true if provider.latest?
+        when :absent
+          return true if is == :absent
+        end
       end
       false
     end
-
   end
 
   # Make sure cpanm installation happens first
@@ -46,21 +43,21 @@ Puppet::Type.newtype(:cpanm) do
     'cpanm'
   end
 
-  newparam(:name, :namevar => true) do
-    desc "The CPAN module to manage"
+  newparam(:name, namevar: true) do
+    desc 'The CPAN module to manage'
   end
 
-  newparam(:force, :boolean => true) do
-    desc "Force installation"
+  newparam(:force, boolean: true) do
+    desc 'Force installation'
     defaultto :false
   end
 
-  newparam(:test, :boolean => true) do
-    desc "Run CPAN module tests"
+  newparam(:test, boolean: true) do
+    desc 'Run CPAN module tests'
     defaultto :false
   end
 
   newparam(:mirror) do
-    desc "URL of the CPAN mirror to use"
+    desc 'URL of the CPAN mirror to use'
   end
 end
